hdkeychain: Update NewMaster to accept network.
This changes the NewMaster function to accept the network the generated
extended master key is associated with.  This could previously be done
by calling SetNet on the returned extended key, but that approach is
more error prone since it is easy for a caller to forget to do it or
never know they should to begin with.