btcutil: update module to v2
In this commit, we update the top-level go module as well as all the
relevant imports to point to the new v2 version of the module. This
change is long overdue after a breaking change was made to the
[`hdkeychain` package](github.com/btcsuite/btcutil) to fix a BIP 32
implementation divergence. A follow up PR will also update the `psbt`
module as well as significant changes have been made to the API and
logic since it was first added.